sync: keep CmuxSyncStore one actor-private file; budget it instead of splitting
Autoreview flagged that splitting the store into a +SQLite extension
forced the nonisolated(unsafe) sqlite3 handle from file-private to
module-internal, weakening the actor-isolation invariant (any future
module file could touch the raw handle off-actor). Revert the store split
to keep `db` private to the actor, and accept the 619-line file as
documented known debt in swift-file-length-budget.tsv (the guard's
explicit escape). The test-file split is kept (pure test code, no
concurrency invariant).
